requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
azure-cognitiveservices-speech==1.34.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
exceptiongroup==1.2.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
humanfriendly==10.0
idna==3.6
imageio==2.33.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.351
langchain-community==0.0.4
langchain-core==0.1.1
langchain-experimental==0.0.47
langsmith==0.0.72
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.9.3.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
onnxruntime==1.15.0
openai==0.28.1
opencv-contrib-python==4.8.1.78
opencv-python-headless==4.7.0.72
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==3.20.3
pyarrow==14.0.2
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pydeck==0.8.1b0
pygments==2.17.2
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywavelets==1.5.0
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scikit-image==0.18.3
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
streamlit==1.29.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2023.12.9
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
yarl==1.9.4
zipp==3.17.0